% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkdata.R
\name{mkdata}
\alias{mkdata}
\title{Return the data of gender and age based on the ratio of the population of Japan}
\usage{
mkdata(n)
}
\arguments{
\item{n}{number of records}
}
\value{
dataframe
}
\description{
Return the data of gender and age based on the ratio of the population of Japan
}
