% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synJP.R
\name{synJP}
\alias{synJP}
\title{Generate synthsetic data from samples based on Japanese statistics}
\usage{
synJP(V, n, name = FALSE)
}
\arguments{
\item{V}{the name of variebles generate}

\item{n}{number of records}

\item{name}{logical value with generating name's data}
}
\value{
character
}
\description{
Generate synthsetic data from samples based on Japanese statistics
}
