% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synJP.R
\name{synJP}
\alias{synJP}
\title{JPdata→synthpop→JPdata2(no name)}
\usage{
synJP(V, n, name = FALSE)
}
\arguments{
\item{V}{variebles' name}

\item{n}{number of records}

\item{name}{make name's data or not}
}
\value{
character
}
\description{
hoge
}
